requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
PyMySQL==1.1.1
stripe==10.6.0
python-dotenv==1.0.1
APScheduler==3.10.4
